refactor: Scaffold SignUpScreen 내부로 넣기
