+++ app/src/androidTest/java/nextstep/signup/component/InformationTextFieldTest.kt
@@ -14,13 +14,13 @@ import androidx.compose.ui.test.performTextInput
 import junit.framework.Assert.assertEquals
 import junit.framework.Assert.assertFalse
 import junit.framework.Assert.assertTrue
-import nextstep.signup.ui.component.SignUpTextField
+import nextstep.signup.ui.component.InformationTextField
 import nextstep.signup.ui.model.UserName
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
-class SignUpTextFieldTest {
+class InformationTextFieldTest {
     @get:Rule
     val composeTestRule = createComposeRule()
 
@@ -33,7 +33,7 @@ class SignUpTextFieldTest {
     @Before
     fun setUp() {
         composeTestRule.setContent {
-            SignUpTextField(
+            InformationTextField(
                 modifier = Modifier.testTag(tag),
                 label = label,
                 value = userName.text,
