+++ app/src/androidTest/java/nextstep/signup/component/SubmitButtonTest.kt
@@ -16,7 +16,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
-class SignUpButtonTest {
+class SubmitButtonTest {
     @get:Rule
     val composeTestRule = createComposeRule()
